Adding validation and daemonized mode for notebooks

+++ jupyterhub_client/execute.py
@@ -2,48 +2,40 @@ import uuid
 import difflib
 import logging
 import textwrap
+import datetime
 
 from jupyterhub_client.api import JupyterHubAPI, JupyterKernelAPI, JupyterAPI
-from jupyterhub_client.utils import parse_notebook_cells
+from jupyterhub_client.utils import parse_notebook_cells, tangle_cells
 
 logger = logging.getLogger(__name__)
 
 
-async def execute_code(hub_url, cells, username=None, create_user=False, delete_user=False, username_format='user-{id}', timeout=None):
+async def execute_code(hub_url, cells, username=None, create_user=False, delete_user=False, username_format='user-{id}', timeout=None, daemonized=False, validate=False):
     username = username or username_format.format(id=str(uuid.uuid4()))
     hub = JupyterHubAPI(hub_url)
 
     async with hub:
         try:
-            user = await hub.get_user(username)
-            if user is None:
-                if create_user:
-                    await hub.create_user(username)
-                else:
-                    raise ValueError(f'current username={username} does not exist and create_user={create_user}')
-                user = await hub.get_user(username)
-
-            if user['server'] is None:
-                await hub.create_server(username)
-
-            jupyter = JupyterAPI(hub.hub_url / 'user' / username, hub.api_token)
+            jupyter = await hub.ensure_server(username, create_user=create_user)
 
             async with jupyter:
-                kernel_id = (await jupyter.create_kernel())['id']
-                kernel = JupyterKernelAPI(jupyter.api_url / 'kernels' / kernel_id, jupyter.api_token)
+                kernel_id, kernel = await jupyter.ensure_kernel()
                 async with kernel:
                     for i, (code, expected_result) in enumerate(cells):
-                        kernel_result = await kernel.send_code(username, code, timeout=timeout)
-                        logger.debug(f'kernel execucting cell={i} code=\n{textwrap.indent(code, "   >>> ")}')
-                        logger.debug(f'kernel result cell={i} result=\n{textwrap.indent(kernel_result, "   | ")}')
-                        if kernel_result != expected_result:
-                            diff = ''.join(difflib.unified_diff(kernel_result, expected_result))
-                            logger.error(f'kernel result did not match expected result diff={diff}')
-                            raise ValueError(f'execution of cell={i} did not match expected result diff={diff}')
-                await jupyter.delete_kernel(kernel_id)
-            await hub.delete_server(username)
+                        kernel_result = await kernel.send_code(username, code, timeout=timeout, wait=(not daemonized))
+                        if not daemonized:
+                            logger.debug(f'kernel execucting cell={i} code=\n{textwrap.indent(code, "   >>> ")}')
+                            logger.debug(f'kernel result cell={i} result=\n{textwrap.indent(kernel_result, "   | ")}')
+                            if validate and kernel_result != expected_result:
+                                diff = ''.join(difflib.unified_diff(kernel_result, expected_result))
+                                logger.error(f'kernel result did not match expected result diff={diff}')
+                                raise ValueError(f'execution of cell={i} did not match expected result diff={diff}')
+                if not daemonized:
+                    await jupyter.delete_kernel(kernel_id)
+            if not daemonized:
+                await hub.delete_server(username)
         finally:
-            if delete_user:
+            if delete_user and not daemonized:
                 await hub.delete_user(username)
